product_service/requirements.txt
asgiref==3.5.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-filter==22.1
django-ip-geolocation==1.6.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
haversine==2.7.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
pluggy==1.0.0
pytest==7.2.0
pytest-django==4.5.2
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
mysqlclient
